+++ src/model/realtime.rs
@@ -5,10 +5,10 @@ use serde::Serialize;
 #[derive(Serialize)]
 pub struct CurrentPlayer<'a> {
     pub damaging_abilities: Vec<&'static str>,
-    pub damaging_items: Vec<usize>,
-    pub damaging_runes: Vec<usize>,
+    pub damaging_items: Vec<u32>,
+    pub damaging_runes: Vec<u32>,
     pub riot_id: &'a str,
-    pub level: usize,
+    pub level: u8,
     pub team: &'a str,
     pub position: &'a str,
     pub champion_name: &'a str,
@@ -20,8 +20,8 @@ pub struct CurrentPlayer<'a> {
 
 #[derive(Serialize)]
 pub struct GameInformation {
-    pub game_time: f64,
-    pub map_number: usize,
+    pub game_time: f32,
+    pub map_number: u8,
 }
 
 #[derive(Serialize)]
@@ -29,7 +29,7 @@ pub struct Enemy<'a> {
     pub champion_name: &'a str,
     pub riot_id: &'a str,
     pub team: &'a str,
-    pub level: usize,
+    pub level: u8,
     pub position: &'a str,
     pub damages: Damages,
     pub base_stats: BasicStats,
@@ -41,10 +41,10 @@ pub struct Enemy<'a> {
 
 #[derive(Serialize)]
 pub struct Scoreboard<'a> {
-    pub assists: usize,
-    pub creep_score: usize,
-    pub deaths: usize,
-    pub kills: usize,
+    pub assists: u16,
+    pub creep_score: u16,
+    pub deaths: u16,
+    pub kills: u16,
     pub riot_id: &'a str,
     pub champion_id: &'static str,
     pub champion_name: &'a str,
@@ -56,7 +56,7 @@ pub struct Realtime<'a> {
     pub current_player: CurrentPlayer<'a>,
     pub enemies: FxHashMap<&'static str, Enemy<'a>>,
     pub game_information: GameInformation,
-    pub recommended_items: &'static [usize],
+    pub recommended_items: &'static [u32],
     pub scoreboard: Scoreboard<'a>,
     pub enemy_dragon_multipliers: DragonMultipliers,
     pub ally_dragon_multipliers: DragonMultipliers,
